+++ src/parser/tokenizers/commodity.rs
@@ -19,9 +19,9 @@ impl<'a> Tokenizer<'a> {
 
         while let Some(part) = parsed.next() {
             match part.as_rule() {
-                Rule::comment => comments.push(Comment {
-                    comment: parse_string(part),
-                }),
+                Rule::comment => comments.push(Comment::from(parse_string(
+                    part.into_inner().next().unwrap(),
+                ))),
                 Rule::commodity_property => {
                     let mut property = part.into_inner();
                     match property.next().unwrap().as_rule() {