+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 
+## [1.0.0] - planned
+### Added
+- Support for specifying payees via posting comments.
+- Added support for dates in posting comments
+### Changed
+- Date comparisons are done at the posting level rather than the transaction level
 ## [0.17.0] - 2021-03-12
 ### Changed
 - Now the whole file is processed using a formal grammar